chore(rename): folder, naming, paths
