+++ docs/.vuepress/config.ts
@@ -19,7 +19,7 @@ export default defineUserConfig({
   alias: {
     // Alias src directory for @/components/{KongponentName} imports
     '@': path.resolve(__dirname, '../../src'),
-    utils: path.resolve(__dirname, '../../utils'),
+    mocks: path.resolve(__dirname, '../../mocks'),
   },
   // when using vuepress-vite package, you can omit this field
   // because vite is the default bundler
